requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.29
protobuf==4.21.2
